+++ anndata/core/anndata.py
@@ -23,6 +23,9 @@ from pandas.api.types import is_string_dtype, is_categorical
 from scipy import sparse
 from scipy.sparse import issparse
 
+from .raw import Raw
+from .index import _normalize_indices, Index, Index1D
+from .file_backing import AnnDataFileManager
 from .alignedmapping import (
     AxisArrays,
     AxisArraysView,
@@ -42,13 +45,7 @@ from .views import (
 )
 from .sparsedataset import SparseDataset
 from .. import utils
-from ..utils import (
-    Index1D,
-    Index,
-    convert_to_dict,
-    unpack_index,
-    ensure_df_homogeneous,
-)
+from ..utils import convert_to_dict, ensure_df_homogeneous
 from ..logging import anndata_logger as logger
 from ..compat import ZarrArray, ZappyArray, DaskArray, Literal
 
@@ -94,59 +91,6 @@ def _check_2d_shape(X):
         )
 
 
-def _normalize_index(
-    indexer, index: pd.Index
-) -> Union[slice, int, "np.ndarray[int]"]:
-    if not isinstance(index, RangeIndex):
-        assert (
-            index.dtype != float and index.dtype != int
-        ), 'Don’t call _normalize_index with non-categorical/string names'
-
-    # the following is insanely slow for sequences, we replaced it using pandas below
-    def name_idx(i):
-        if isinstance(i, str):
-            i = index.get_loc(i)
-        return i
-
-    if isinstance(indexer, slice):
-        start = name_idx(indexer.start)
-        stop = name_idx(indexer.stop)
-        # string slices can only be inclusive, so +1 in that case
-        if isinstance(indexer.stop, str):
-            stop = None if stop is None else stop + 1
-        step = indexer.step
-        return slice(start, stop, step)
-    elif isinstance(indexer, (np.integer, int)):
-        return indexer
-    elif isinstance(indexer, str):
-        return index.get_loc(indexer)  # int
-    elif isinstance(indexer, (Sequence, np.ndarray, pd.Index)):
-        if not isinstance(indexer, (np.ndarray, pd.Index)):
-            indexer = np.array(indexer)
-        if issubclass(indexer.dtype.type, (np.integer, np.floating)):
-            return indexer  # Might not work for range indexes
-        elif issubclass(indexer.dtype.type, np.bool_):
-            if indexer.shape != index.shape:
-                raise IndexError(
-                    f"Boolean index does not match AnnData's shape along this "
-                    f"dimension. Boolean index has shape {indexer.shape} while "
-                    f"AnnData index has shape {index.shape}."
-                )
-            positions = np.where(indexer)[0]
-            return positions  # np.ndarray[int]
-        else:  # indexer should be string array
-            positions = index.get_indexer(indexer)
-            if np.any(positions < 0):
-                not_found = indexer[positions < 0]
-                raise KeyError(
-                    f"Values {list(not_found)}, from {list(indexer)}, "
-                    "are not valid obs/ var names or indices."
-                )
-            return positions  # np.ndarray[int]
-    else:
-        raise IndexError(f'Unknown indexer {indexer!r} of type {type(indexer)}')
-
-
 def _gen_dataframe(anno, length, index_names):
     if isinstance(anno, pd.DataFrame):
         anno = anno.copy()
@@ -170,293 +114,6 @@ def _gen_dataframe(anno, length, index_names):
                 anno, index=RangeIndex(0, length, name=None).astype(str)
             )
     return _anno
-
-
-class AnnDataFileManager:
-    """Backing file manager for AnnData."""
-
-    def __init__(
-        self,
-        adata: 'AnnData',
-        filename: Optional[PathLike] = None,
-        filemode: Optional[Literal['r', 'r+']] = None,
-    ):
-        self._adata = adata
-        self.filename = filename
-        self._filemode = filemode
-        self._file = None
-        if filename:
-            self.open()
-
-    def __repr__(self) -> str:
-        if self.filename is None:
-            return 'Backing file manager: no file is set.'
-        else:
-            return f'Backing file manager of file {self.filename}.'
-
-    def __contains__(self, x) -> bool:
-        return x in self._file
-
-    def __getitem__(
-        self, key: str
-    ) -> Union[h5py.Group, h5py.Dataset, SparseDataset]:
-        return self._file[key]
-
-    def __setitem__(
-        self, key: str, value: Union[h5py.Group, h5py.Dataset, SparseDataset]
-    ):
-        self._file[key] = value
-
-    def __delitem__(self, key: str):
-        del self._file[key]
-
-    @property
-    def filename(self) -> Path:
-        return self._filename
-
-    @filename.setter
-    def filename(self, filename: Optional[PathLike]):
-        self._filename = None if filename is None else Path(filename)
-
-    def open(
-        self,
-        filename: Optional[PathLike] = None,
-        filemode: Optional[Literal['r', 'r+']] = None,
-    ):
-        if filename is not None:
-            self.filename = filename
-        if filemode is not None:
-            self._filemode = filemode
-        if self.filename is None:
-            raise ValueError(
-                'Cannot open backing file if backing not initialized.'
-            )
-        self._file = h5py.File(self.filename, self._filemode)
-
-    def close(self):
-        """Close the backing file, remember filename, do *not* change to memory mode."""
-        if self._file is not None:
-            self._file.close()
-
-    def _to_memory_mode(self):
-        """Close the backing file, forget filename, *do* change to memory mode."""
-        self._adata.__X = self._adata.X[()]
-        self._file.close()
-        self._file = None
-        self._filename = None
-
-    @property
-    def isopen(self) -> bool:
-        """State of backing file."""
-        if self._file is None:
-            return False
-        # try accessing the id attribute to see if the file is open
-        return bool(self._file.id)
-
-
-# TODO: Implement views for Raw
-class Raw:
-    def __init__(
-        self,
-        adata: Optional['AnnData'] = None,
-        X: Union[np.ndarray, sparse.spmatrix, None] = None,
-        var: Union[AxisArrays, AxisArraysView, None] = None,
-        varm: Union[AxisArrays, AxisArraysView, None] = None,
-    ):
-        self._adata = adata
-        self._n_obs = adata.n_obs
-        if X is not None:
-            self._X = X
-            self._var = var
-            self._varm = AxisArrays(self, 1, varm)
-        else:
-            self._X = None if adata.isbacked else adata.X.copy()
-            self._var = adata.var.copy()
-            self._varm = AxisArrays(self, 1, adata.varm.copy())
-
-    @property
-    def X(self):
-        # TODO: Handle unsorted array of integer indices for h5py.Datasets
-        if self._adata.isbacked:
-            if not self._adata.file.isopen:
-                self._adata.file.open()
-            # Handle legacy file formats:
-            if "raw/X" in self._adata.file:
-                X = self._adata.file["raw/X"]
-            elif "raw.X" in self._adata.file:
-                X = self._adata.file['raw.X']  # Backwards compat
-            else:
-                raise AttributeError(
-                    f"Could not find dataset for raw X in file: "
-                    f"{self._adata.file.filename}."
-                )
-            if isinstance(X, h5py.Group):
-                X = SparseDataset(X)
-            # Check if we need to subset
-            if self._adata.isview:
-                # TODO: As noted above, implement views of raw
-                #       so we can know if we need to subset by var
-                return X[self._adata._oidx, slice(None)]
-            else:
-                return X
-        else:
-            return self._X
-
-    @property
-    def shape(self):
-        return self.n_obs, self.n_vars
-
-    @property
-    def var(self):
-        return self._var
-
-    @property
-    def n_vars(self):
-        return self._var.shape[0]
-
-    @property
-    def n_obs(self):
-        return self._n_obs
-
-    @property
-    def varm(self):
-        return self._varm
-
-    @property
-    def var_names(self):
-        return self.var.index
-
-    @property
-    def obs_names(self):
-        return self._adata.obs_names
-
-    def __getitem__(self, index):
-        oidx, vidx = self._normalize_indices(index)
-
-        # To preserve two dimensional shape
-        if isinstance(vidx, (int, np.integer)):
-            vidx = slice(vidx, vidx + 1, 1)
-        if isinstance(oidx, (int, np.integer)):
-            oidx = slice(oidx, oidx + 1, 1)
-
-        if not self._adata.isbacked:
-            X = _subset(self.X, (oidx, vidx))
-        else:
-            X = None
-
-        var = self._var.iloc[vidx]
-        new = Raw(self._adata, X=X, var=var)
-        if self._varm is not None:
-            new._varm = self._varm._view(
-                self, (vidx,)
-            ).copy()  # Since there is no view of raws
-        return new
-
-    def copy(self):
-        return Raw(
-            self._adata,
-            X=self._X.copy(),
-            var=self._var.copy(),
-            varm=None if self._varm is None else self._varm.copy(),
-        )
-
-    def to_adata(self):
-        """Create full AnnData object.
-        """
-        return AnnData(
-            X=self._X.copy(),
-            var=self._var.copy(),
-            varm=None if self._varm is None else self._varm.copy(),
-            obs=self._adata.obs.copy(),
-            obsm=self._adata.obsm.copy(),
-            uns=self._adata.uns.copy(),
-        )
-
-    def _normalize_indices(self, packed_index):
-        # deal with slicing with pd.Series
-        if isinstance(packed_index, pd.Series):
-            packed_index = packed_index.values
-        if isinstance(packed_index, tuple):
-            if len(packed_index) != 2:
-                raise IndexDimError(len(packed_index))
-            if isinstance(packed_index[1], pd.Series):
-                packed_index = packed_index[0], packed_index[1].values
-            if isinstance(packed_index[0], pd.Series):
-                packed_index = packed_index[0].values, packed_index[1]
-        obs, var = unpack_index(packed_index)
-        obs = _normalize_index(obs, self._adata.obs_names)
-        var = _normalize_index(var, self.var_names)
-        return obs, var
-
-    def var_vector(self, k: str) -> np.ndarray:
-        """\
-        Convenience function for returning a 1 dimensional ndarray of values
-        from `.X` or `.var`.
-
-        Made for convenience, not performance. Intentionally permissive about
-        arguments, for easy iterative use.
-
-        Params
-        ------
-        k
-            Key to use. Should be in `.obs_names` or `.var.columns`.
-
-        Returns
-        -------
-        A one dimensional nd array, with values for each var in the same order
-        as `.var_names`.
-        """
-        if k in self.var:
-            return self.var[k].values
-        else:
-            idx = self._normalize_indices((k, slice(None)))
-            a = self.X[idx]
-        if issparse(a):
-            a = a.toarray()
-        return np.ravel(a)
-
-    def obs_vector(self, k: str) -> np.ndarray:
-        """\
-        Convenience function for returning a 1 dimensional ndarray of values
-        from `.X`.
-
-        Made for convenience, not performance. Intentionally permissive about
-        arguments, for easy iterative use.
-
-        Params
-        ------
-        k
-            Key to use. Should be in `.var_names` or `.obs.columns`. If `use_raw`,
-            value should be in `.raw.var_names` instead of `.var_names`.
-
-        Returns
-        -------
-        A one dimensional nd array, with values for each obs in the same order
-        as `.obs_names`.
-        """
-        idx = self._normalize_indices((slice(None), k))
-        a = self.X[idx]
-        if issparse(a):
-            a = a.toarray()
-        return np.ravel(a)
-
-
-INDEX_DIM_ERROR_MSG = (
-    'You tried to slice an AnnData(View) object with an'
-    '{}-dimensional index, but only 2 dimensions exist in such an object.'
-)
-INDEX_DIM_ERROR_MSG_1D = (
-    '\nIf you tried to slice cells using adata[cells, ], '
-    'be aware that Python (unlike R) uses adata[cells, :] as slicing syntax.'
-)
-
-
-class IndexDimError(IndexError):
-    def __init__(self, n_dims):
-        msg = INDEX_DIM_ERROR_MSG.format(n_dims)
-        if n_dims == 1:
-            msg += INDEX_DIM_ERROR_MSG_1D
-        super().__init__(msg)
 
 
 class ImplicitModificationWarning(UserWarning):
@@ -838,12 +495,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         elif isinstance(raw, Raw):
             self._raw = raw
         else:
-            # is dictionary from reading the file, nothing that is meant for a user
+            # is dictionary from reading the file,
+            # nothing that is meant for a user
             if self.isbacked:
                 raw_key = "raw/X" if "raw/X" in self.file else "raw.X"
                 if raw_key not in self.file:
                     raise KeyError(
-                        f"Tried to check shape of raw from {self.file.filename}, but there was no entry for 'raw/X'."
+                        f"Tried to check shape of raw from {self.filename}, "
+                        f"but there was no entry for 'raw/X'."
                     )
                 if isinstance(self.file[raw_key], h5py.Group):
                     shape = self.file[raw_key].attrs["shape"]
@@ -913,7 +572,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     def X(self) -> Optional[Union[np.ndarray, sparse.spmatrix, ArrayView]]:
         """Data matrix of shape :attr:`n_obs` × :attr:`n_vars`."""
         if self.isbacked:
-            if not self.file.isopen:
+            if not self.file.is_open:
                 self.file.open()
             X = self.file['X']
             if isinstance(X, h5py.Group):
@@ -970,7 +629,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 raise ValueError('Not implemented.')
             self._X = None
             return
-        # If indices are both arrays, we need to modify them so we don't set values like coordinates
+        # If indices are both arrays, we need to modify them
+        # so we don’t set values like coordinates
         # This can occur if there are succesive views
         if (
             self.isview
@@ -1009,7 +669,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                     self._X = value
         else:
             raise ValueError(
-                f'Data matrix has wrong shape {value.shape}, need to be {self.shape}.'
+                f'Data matrix has wrong shape {value.shape}, '
+                f'need to be {self.shape}.'
             )
 
     @property
@@ -1317,7 +978,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return self._isview
 
     @property
-    def filename(self) -> Optional[PathLike]:
+    def filename(self) -> Optional[Path]:
         """\
         Change to backing mode by setting the filename of a ``.h5ad`` file.
 
@@ -1365,27 +1026,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         write_attribute(self.file._file, attr, value)
 
     def _normalize_indices(self, index: Optional[Index]) -> Tuple[slice, slice]:
-        # deal with tuples of length 1
-        if isinstance(index, tuple) and len(index) == 1:
-            index = index[0]
-        # deal with pd.Series
-        if isinstance(index, pd.Series):
-            index: Index = index.values
-        if isinstance(index, tuple):
-            if len(index) > 2:
-                raise ValueError(
-                    'AnnData can only be sliced in rows and columns.'
-                )
-            # deal with pd.Series
-            # TODO: The series should probably be aligned first
-            if isinstance(index[1], pd.Series):
-                index = index[0], index[1].values
-            if isinstance(index[0], pd.Series):
-                index = index[0].values, index[1]
-        obs, var = unpack_index(index)
-        obs = _normalize_index(obs, self.obs_names)
-        var = _normalize_index(var, self.var_names)
-        return obs, var
+        return _normalize_indices(index, self.obs_names, self.var_names)
 
     # TODO: this is not quite complete...
     def __delitem__(self, index: Index):
